Added Debug trait for print debugging to all structs. Midi parsing and printing Notes with correct frequency and duration but possilby bugs in volume and definitely in start_time.

+++ src/audio_gen.rs
@@ -46,7 +46,7 @@ where
         None
     ).unwrap();
     stream.play().unwrap();
-    std::thread::sleep(time::Duration::from_millis(note.duration_ms));
+    std::thread::sleep(time::Duration::from_millis(note.duration_ms as u64));
 }
 
 fn gen_notes_impl<T>(device: &cpal::Device, config: &cpal::StreamConfig,
